+++ Source/Mosa.Compiler.Framework/Transform/BaseTransformation.cs
@@ -767,7 +767,8 @@ namespace Mosa.Compiler.Framework.Transform
 
 		#region Status Helpers
 
-		public enum TriState { Yes, No, Unknown };
+		public enum TriState
+		{ Yes, No, Unknown };
 
 		public static TriState AreAnyStatusFlagsUsed(Context context)
 		{
@@ -852,8 +853,6 @@ namespace Mosa.Compiler.Framework.Transform
 
 		protected static void RemoveRestOfInstructions(Context context)
 		{
-			var block = context.Block;
-
 			context.GotoNext();
 
 			while (!context.IsBlockEndInstruction)
