+++ Source/Mosa.Compiler.Framework/Transform/Manual/ManualTransforms.cs
@@ -31,6 +31,7 @@ namespace Mosa.Compiler.Framework.Transform.Manual
 
 			new ConstantFolding.Branch32(),
 			new ConstantFolding.Branch64(),
+			new ConstantFolding.Switch(),
 
 			new Special.Deadcode(),
 			new Special.GetLow32From64(),
